[moduledependency] Rename test class to SnomedModuleDependencyRefsetTest
And placed into the components to the snomed.api.rest.components package
