+++ snomed/com.b2international.snowowl.snomed.api.rest.tests/src/com/b2international/snowowl/snomed/api/rest/AllSnomedApiTests.java
@@ -36,6 +36,7 @@ import com.b2international.snowowl.snomed.api.rest.browser.SnomedBrowserApiTest;
 import com.b2international.snowowl.snomed.api.rest.classification.SnomedClassificationApiTest;
 import com.b2international.snowowl.snomed.api.rest.components.SnomedConceptApiTest;
 import com.b2international.snowowl.snomed.api.rest.components.SnomedDescriptionApiTest;
+import com.b2international.snowowl.snomed.api.rest.components.SnomedModuleDependencyRefsetTest;
 import com.b2international.snowowl.snomed.api.rest.components.SnomedRefSetApiTest;
 import com.b2international.snowowl.snomed.api.rest.components.SnomedRefSetBulkApiTest;
 import com.b2international.snowowl.snomed.api.rest.components.SnomedRefSetMemberApiTest;
@@ -48,7 +49,6 @@ import com.b2international.snowowl.snomed.api.rest.ext.SnomedExtensionVersioning
 import com.b2international.snowowl.snomed.api.rest.id.SnomedIdentifierApiTest;
 import com.b2international.snowowl.snomed.api.rest.io.SnomedExportApiTest;
 import com.b2international.snowowl.snomed.api.rest.io.SnomedImportApiTest;
-import com.b2international.snowowl.snomed.api.rest.moduledependency.SnomedModuleDependencyApiTest;
 import com.b2international.snowowl.snomed.api.rest.perf.SnomedConceptCreatePerformanceTest;
 import com.b2international.snowowl.snomed.api.rest.versioning.SnomedVersioningApiTest;
 import com.b2international.snowowl.snomed.common.SnomedTerminologyComponentConstants;
@@ -86,7 +86,7 @@ import com.b2international.snowowl.test.commons.SnowOwlAppRule;
 	SnomedClassificationApiTest.class,
 	SnomedExportApiTest.class,
 	// Module dependecy test cases
-	SnomedModuleDependencyApiTest.class,
+	SnomedModuleDependencyRefsetTest.class,
 	// Extension test cases
 	SnomedExtensionUpgradeTest.class,
 	SnomedExtensionDowngradeTest.class,
